Added class RectangularArray to represent a rectangular array of elements on a plane.
This class can be used to represent pixels in an image sensor (for
example); in this case, the RectangularArray's regular structure can be
utilized to determine which pixels intersect with an image.

RectangularArray has been written with property access (set) methods to
check the values of the RectangularArray properties.

Method normal provides the normal vector.

Method listIntersectingElements lists the indices of the elements of the
RectangularArray that intersect a given polygon.  In its current (simple)
implementation, it also returns the indices of some array elements that do
NOT intersect the given polygon.

Method listIntersectingElements uses method
convertGlobalToArrayCoordinates.