Replaced properties axis_horizontal and axis_vertical in class RectangularArray with a OrthogonalAxes object.
This change facilitates error checking to ensure that the two axes are
perpendicular to each other.  Previously, the method
RectangularArray.setAxes was used so that both the axes can be
simultaneously updated (so that the orthogonality check doesn't fail after
changing one axis), but this required copying and replacing the
RectangularArray object to make the change.

Created class OrthogonalAxes.

Moved method normal from RectangularArray to OrthogonalAxes.