Revised tests in RectangularArrayTest.
Modified testConstructorHappy to verify that the properties of the
RectangularArray object are properly set.

Created testConstructorDefaultPolygon to ensure that the properties are
still properly set, even if no polygon_template is specified in the
arguments.

Modified testListIntersectingElementsIncludesAll to more appropriately
handle the case where the randomly generated polygon does not intersect
the elements of the RectangularArray.  Now, it will try again with a new
intersecting_polygon (but the same RectangularArray).